a bit refactor, del unused menus
