--- bot/handlers/cabinet_heandlers/__init__.py
@@ -1,4 +1,0 @@
-from .cabinet import router as main_cabinet_router
-
-
-routers = [main_cabinet_router]
