+++ bot/handlers/my_account_heandlers/__init__.py
@@ -1,3 +1,4 @@
 from .my_account import router as my_account_router
+from .referrals import router as referrals_router
 
-routers = [my_account_router]
+routers = [my_account_router, referrals_router]
