+++ bot/handlers/other_handlers/__init__.py
@@ -1,3 +1,4 @@
 from .others import router as others_router
+from .language import router as language_router
 
-routers = [others_router]
+routers = [others_router, language_router]
