--- bot/handlers/setting_heandlers/__init__.py
@@ -1,3 +1,0 @@
-from .setting import router as language_router
-
-routers = [language_router]
